restructure package
