+++ src/enforce_notebook_run_order/__init__.py
@@ -1,0 +1,3 @@
+"""Ensures that jupyter notebooks are run in order"""
+
+from .enforce_notebook_run_order import check_single_notebook, process_path
